requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-crontab==0.7.1
django-filter==24.3
djangorestframework==3.15.2
sqlparse==0.5.1
